moves queryrow items to their own directory
