+++ src/app/history-query.component.ts
@@ -6,7 +6,7 @@
 import { ChangeDetectorRef, Component, Input, OnInit } from '@angular/core';
 import { ISampleQuery } from './base';
 import { QueryRunnerService } from './query-runner.service';
-import { QueryRowComponent } from './queryrow.component';
+import { QueryRowComponent } from './queryrow/queryrow.component';
 
 declare let moment: any;
