+++ src/app/queryrow/queryrow.component.css
@@ -65,3 +65,4 @@ method-badge {
     margin:auto;
     height: 22px;
 }
+
